fix malformed MIME header line error
The headers and body must be separated by a blank line. Strict and/or
brittle servers may fail to parse the headers, resulting in an error:
"malformed MIME header line"

This adds a newline between the headers and message body.

A test is also added to ensure the net/mail std lib can parse the
message.